feat: added temp rbac dashboard -3
